+++ src/Characters/Components/HeroAnimComp.h
@@ -1,10 +1,9 @@
 #pragma once
 #include <SFML/Graphics.hpp>
-#include "../../Managers/StateEnums.h"
 #include "../../Core/Components/BaseAnimComp.h"
+
 namespace ETG
 {
-    enum class HeroStateEnum;
     class Hero;
 
     class HeroAnimComp : public BaseAnimComp<HeroStateEnum>
@@ -15,9 +14,9 @@ namespace ETG
         //Override
         void Update() override;
         void SetAnimations() override;
-        
-        template<typename... TObjects>
-         sf::Vector2f FlipSprites(const Direction& currentDirection, TObjects&... objects);
+
+        template <typename... TObjects>
+        sf::Vector2f FlipSprites(const Direction& currentDirection, TObjects&... objects);
 
     private:
         Hero* HeroPtr = nullptr;
