+++ src/Characters/Hero.h
@@ -2,15 +2,16 @@
 #include <SFML/System/Vector2.hpp>
 #include <SFML/Graphics.hpp>
 #include "../Core/GameObject.h"
-#include "../Guns/RogueSpecial/RogueSpecial.h"
 #include "../Managers/StateEnums.h"
-#include "Components/HeroMoveComp.h"
-#include "Components/InputComponent.h"
-#include "../Characters/Components/HeroAnimComp.h"
+#include "../Core/Factory.h"
 
 namespace ETG
 {
-
+    class RogueSpecial;
+    class HeroAnimComp;
+    class InputComponent;
+    class HeroMoveComp;
+    
     class Hero : public GameObject
     {
     public:
@@ -19,6 +20,7 @@ namespace ETG
         static bool IsShooting;
 
         explicit Hero(sf::Vector2f Position);
+        ~Hero() override;
         void Update() override;
         void Draw() override;
         void DrawHeroBounds() const;
@@ -32,7 +34,7 @@ namespace ETG
 
     private:
         std::unique_ptr<HeroAnimComp> AnimationComp;
-        InputComponent InputComp;
+        std::unique_ptr<InputComponent> InputComp;
 
         sf::Texture HandTex;
         sf::Vector2f HandPos;
