+++ src/Managers/GameManager.h
@@ -1,23 +1,34 @@
 #pragma once
 
-#include "DebugTexts.h"
-#include "../Animation/AnimationManager.h"
+#include "Globals.h"
 #include "../Characters/Hero.h"
 #include "../UI/UserInterface.h"
 #include "../Engine/Engine.h"
 
 namespace ETG
 {
-    using namespace Globals;
+    class DebugText;
 
     class GameManager
     {
+    public:
+        GameManager();
+        ~GameManager();
+        void Initialize();
+        void ProcessEvents();
+        [[nodiscard]] bool WindowHasFocus() const { return HasFocus; }
+
+        //I might delete this later on 
+        static bool IsRunning() { return Globals::Window->isOpen(); }
+        void Update();
+        void Draw();
+
     private:
         std::unique_ptr<Hero> Hero;
         Engine EngineUI{};
 
         bool HasFocus = true;
-        DebugText DebugText;
+        std::unique_ptr<DebugText> DebugText;
 
     public:
         //Hold only scene objects. Used for displaying details panel
@@ -25,15 +36,5 @@ namespace ETG
         
         std::unique_ptr<UserInterface> UI;
         static sf::Event GameEvent;
-
-
-        void Initialize();
-        void ProcessEvents();
-        [[nodiscard]] bool WindowHasFocus() const { return HasFocus; }
-
-        //I might delete this later on 
-        static bool IsRunning() { return Window->isOpen(); }
-        void Update();
-        void Draw();
     };
 }
