+++ src/Managers/InputManager.cpp
@@ -1,5 +1,5 @@
 #include "InputManager.h"
-
+#include <complex>
 #include "GameManager.h"
 #include "GameState.h"
 #include "../Engine/Engine.h"
